Refactor authentication components and styles
- Deleted unused AuthMessage, PasswordInput, and related styles.
- Updated ResetPassword and SignIn pages to use new shared auth components.
- Created new AuthButton, AuthInput, AuthLink, and AuthMessage components with improved styling and functionality.
- Consolidated shared components under a new directory structure for better organization.
- Added responsive styles for improved mobile experience.
