--- frontend/src/components/SwitchBox/SwitchBox.css
@@ -1,44 +1,0 @@
-.switch-box {
-  position: relative;
-  display: flex;
-  background-color: gray;
-  height: 1.5rem;
-  width: 2.8rem;
-  border-radius: 1rem;
-  padding: 0.2rem;
-  margin: 0.8rem;
-  transition: left 0.3s ease;
-  cursor: pointer;
-  user-select: none;
-  align-items: center;
-  overflow: visible;
-}
-
-.switch-box.on {
-  background-color: #3b82f6;
-}
-
-.switch-box__ball {
-  width: 1rem;
-  height: 1rem;
-  border-radius: 50%;
-  background-color: rgb(255, 255, 255);
-  position: absolute;
-  transition: background-color 0.3s ease, left 0.3s ease;
-  left: 8%;
-}
-
-.switch-box.on .switch-box__ball {
-  position: absolute;
-  background-color: white;
-  transition: background-color 0.3s ease, left 0.3s ease;
-  left: calc(92% - 1rem);
-}
-
-.switch-box:hover .switch-box__ball {
-  box-shadow: 0 0 15px 5px #b8b8b8;
-}
-
-.switch-box.on:hover .switch-box__ball {
-  box-shadow: 0 0 15px 5px #3b82f6;
-}
