+++ frontend/src/components/shared/auth/AuthButton/AuthButton.tsx
