+++ frontend/src/components/shared/auth/AuthInput/AuthInput.tsx
