+++ frontend/src/components/shared/auth/AuthLink/AuthLink.css
@@ -1,11 +1,11 @@
-@import '../../../styles/auth-variables.css';
+@import "../../../../styles/auth-variables.css";
 
 .auth-link {
   color: var(--auth-primary-color);
   text-decoration: none;
   font-weight: var(--auth-font-semibold);
   transition: color var(--auth-transition-base),
-              transform var(--auth-transition-base);
+    transform var(--auth-transition-base);
   display: inline-flex;
   align-items: center;
   gap: var(--auth-space-xs);
