+++ frontend/src/components/shared/auth/PasswordInput/PasswordInput.css
@@ -1,4 +1,4 @@
-@import '../../../styles/auth-variables.css';
+@import "../../../../styles/auth-variables.css";
 
 .password-input-group {
   display: flex;
@@ -29,9 +29,8 @@
   font-size: 15px;
   color: #1a1a1a;
   background: white;
-  transition: border-color 0.2s ease,
-              background-color 0.2s ease,
-              box-shadow 0.2s ease;
+  transition: border-color 0.2s ease, background-color 0.2s ease,
+    box-shadow 0.2s ease;
   font-family: inherit;
   box-sizing: border-box;
   min-width: 0;
@@ -70,7 +69,7 @@
   cursor: pointer;
   padding: 4px;
   transition: color var(--auth-transition-base),
-              transform var(--auth-transition-base);
+    transform var(--auth-transition-base);
   display: flex;
   align-items: center;
   justify-content: center;
@@ -113,13 +112,13 @@
     font-size: var(--auth-font-base);
     min-height: var(--auth-input-height);
   }
-  
+
   .password-toggle {
     right: var(--auth-space-lg);
     font-size: var(--auth-font-2xl);
     padding: var(--auth-space-sm);
   }
-  
+
   .password-input-label {
     font-size: var(--auth-font-sm);
   }
@@ -131,7 +130,7 @@
     padding: var(--auth-space-md) var(--auth-space-lg);
     padding-right: var(--auth-space-4xl);
   }
-  
+
   .password-toggle {
     right: var(--auth-space-md);
   }
