+++ frontend/src/components/shared/auth/PasswordInput/PasswordInput.tsx
