+++ frontend/src/components/shared/auth/forms/FormSignUp/SignUpForm.tsx
@@ -1,11 +1,14 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { useAuth } from "../../../context/AuthContext";
-import { createSignUpHandler } from "../../../utils/auth";
-import { AuthInput, AuthButton, AuthMessage, PasswordInput } from "../../../components/shared";
-import handleChange from "../../../utils/handleChange";
-import type { SignUpFormData, FormErrors } from "../../../types/auth";
-import "./SignUp.css";
+import { useAuth } from "../../../../../context/AuthContext";
+import { createSignUpHandler } from "../../../../../utils/auth";
+import handleChange from "../../../../../utils/handleChange";
+import type { SignUpFormData, FormErrors } from "../../../../../types/auth";
+import "../../../../../pages/auth/signup/SignUp.css";
+import AuthInput from "../../AuthInput/AuthInput";
+import AuthMessage from "../../AuthMessage/AuthMessage";
+import PasswordInput from "../../PasswordInput/PasswordInput";
+import AuthButton from "../../AuthButton/AuthButton";
 
 export default function SignUpForm() {
   const [formData, setFormData] = useState<SignUpFormData>({
@@ -15,18 +18,16 @@ export default function SignUpForm() {
     password: "",
   });
   const [errors, setErrors] = useState<FormErrors>({});
-  
+
   const { login } = useAuth();
   const navigate = useNavigate();
-  
+
   const handleSubmit = createSignUpHandler(login, navigate, setErrors);
 
   return (
     <form onSubmit={(e) => handleSubmit(e, formData)} className="auth-form">
       {Object.keys(errors).length > 0 && (
-        <AuthMessage type="error">
-          {Object.values(errors)[0]}
-        </AuthMessage>
+        <AuthMessage type="error">{Object.values(errors)[0]}</AuthMessage>
       )}
       <div className="form-row">
         <AuthInput
