+++ frontend/src/components/shared/auth/index.ts
@@ -1,0 +1,7 @@
+import AuthButton from "./AuthButton/AuthButton";
+import AuthInput from "./AuthInput/AuthInput";
+import AuthLink from "./AuthLink/AuthLink";
+import AuthMessage from "./AuthMessage/AuthMessage";
+import PasswordInput from "./PasswordInput/PasswordInput";
+
+export { AuthButton, AuthInput, AuthLink, AuthMessage, PasswordInput };
