--- frontend/src/components/shared/index.ts
@@ -1,6 +1,0 @@
-// Auth components barrel export
-export { default as AuthInput } from './AuthInput/AuthInput';
-export { default as AuthButton } from './AuthButton/AuthButton';
-export { default as AuthMessage } from './AuthMessage/AuthMessage';
-export { default as AuthLink } from './AuthLink/AuthLink';
-export { default as PasswordInput } from './PasswordInput/PasswordInput';
