+++ frontend/src/pages/auth/signin/SignIn.tsx
@@ -1,15 +1,13 @@
-import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { useAuth } from '../../../context/AuthContext';
-import AuthLayout from '../../../components/AuthLayout/AuthLayout';
-import { AuthLink } from '../../../components/shared';
-import { 
-  AUTH_CONSTANTS,
-  createPasswordToggleHandler,
-  createSignInHandler
-} from '../../../utils/auth';
-import './SignIn.css';
-import FormSignIn from './FormSignIn';
+import React, { useState } from "react";
+import { useNavigate } from "react-router-dom";
+import { useAuth } from "../../../context/AuthContext";
+import { createSignInHandler } from "../../../utils/auth";
+import { createPasswordToggleHandler } from "../../../utils/auth";
+import { AUTH_CONSTANTS } from "../../../utils/auth";
+import AuthLayout from "../../../components/AuthLayout/AuthLayout";
+import FormSignIn from "../../../components/shared/auth/forms/FormSignIn/FormSignIn";
+import "./SignIn.css";
+import { AuthLink } from "../../../components/shared/auth";
 
 const SignIn: React.FC = () => {
   const [email, setEmail] = useState("");
@@ -28,16 +26,15 @@ const SignIn: React.FC = () => {
       <div className="signin-container">
         <h1 className="signin-title">Sign in to your account</h1>
         <p className="signin-subtitle">
-          Don't have an account?{' '}
-          <AuthLink to="/auth/sign-up"> 
-            Get started
-          </AuthLink>
+          Don't have an account?{" "}
+          <AuthLink to="/auth/sign-up">Get started</AuthLink>
         </p>
 
         <div className="demo-info">
           <div className="info-icon">ℹ️</div>
           <div className="info-text">
-            Use <strong>{AUTH_CONSTANTS.DEMO_EMAIL}</strong> with password <strong>{AUTH_CONSTANTS.DEMO_PASSWORD}</strong>
+            Use <strong>{AUTH_CONSTANTS.DEMO_EMAIL}</strong> with password{" "}
+            <strong>{AUTH_CONSTANTS.DEMO_PASSWORD}</strong>
           </div>
         </div>
