+++ frontend/src/pages/auth/signup/SignUp.css
@@ -1,5 +1,3 @@
-@import '../../../styles/auth-variables.css';
-
 @keyframes fadeInUp {
   from {
     opacity: 0;
@@ -90,7 +88,7 @@
   .form-header p {
     font-size: var(--auth-font-lg);
   }
-  
+
   .form-row {
     gap: var(--auth-space-md);
   }
@@ -117,7 +115,7 @@
     font-size: var(--auth-font-base);
     margin-bottom: var(--auth-space-xl);
   }
-  
+
   .auth-form {
     gap: var(--auth-form-gap);
   }
@@ -151,4 +149,3 @@
     font-size: var(--auth-font-xs);
   }
 }
-
