+++ frontend/src/pages/auth/signup/SignUp.tsx
@@ -1,7 +1,7 @@
 import { type ReactElement } from "react";
 import AuthLayout from "../../../components/AuthLayout/AuthLayout";
-import { AuthLink } from "../../../components/shared";
-import SignUpForm from "./SignUpForm";
+import { AuthLink } from "../../../components/shared/auth";
+import SignUpForm from "../../../components/shared/auth/forms/FormSignUp/SignUpForm";
 
 export default function SignUp(): ReactElement {
   return (
@@ -11,17 +11,14 @@ export default function SignUp(): ReactElement {
           <h1>Get started absolutely free</h1>
           <p>
             Already have an account?{" "}
-            <AuthLink to="/auth/sign-in">
-              Sign in
-            </AuthLink>
+            <AuthLink to="/auth/sign-in">Sign in</AuthLink>
           </p>
         </div>
 
         <SignUpForm />
 
         <p className="terms-text">
-          By signing up, I agree to{" "}
-          <a href="#">Terms of service</a> and{" "}
+          By signing up, I agree to <a href="#">Terms of service</a> and{" "}
           <a href="#">Privacy policy</a>.
         </p>
       </div>
